refactoring. add voted controller concern
